Rewrite for Canon P23-DH V
Instead of tapping into the existing calculator circuit
and trying really hard not to interfere with its operation,
while counting pulses when a button is pressed; we instead
usurp half the surface area of every button and construct
our own matrix on that half.

Then, the pulse counting and all other garbage becomes
unnecessary. We only need to debounce the presses and
the releases (cf, sleep(30) timeouts).